[IMP] Standardize English flavor
... Or should I say flavour? 🤡 

Recently this subject arose in https://github.com/OCA/sale-workflow/pull/2243#discussion_r1009251042.

In general terms, Odoo core is written in `en_US` and that's the base language for all translations. Since it has some differences with other locales, it's good to have a guideline on what's the correct standard.

I'm not saying we all have to go and change all words in non-American English out there. It's probably too late and it's impractical to do so. However, for new stuff coming, I think it's good to have a standard.